chore(@lexical/text): Split into several files for better readability and to simplify further expansion
It's time to do some housekeeping if we want to add more helpers there :)
