+++ packages/lexical-text/src/canShowPlaceholder.ts
@@ -1,0 +1,78 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+import {
+  $getRoot,
+  $isDecoratorNode,
+  $isElementNode,
+  $isParagraphNode,
+  $isTextNode,
+} from 'lexical';
+
+import {$isRootTextContentEmpty} from './isRootTextContentEmpty';
+
+/**
+ * Determines if the input should show the placeholder. If anything is in
+ * in the root the placeholder should not be shown.
+ * @param isComposing - Is the editor in composition mode due to an active Input Method Editor?
+ * @returns true if the input should show the placeholder, false otherwise.
+ */
+export function $canShowPlaceholder(isComposing: boolean): boolean {
+  if (!$isRootTextContentEmpty(isComposing, false)) {
+    return false;
+  }
+
+  const root = $getRoot();
+  const children = root.getChildren();
+  const childrenLength = children.length;
+
+  if (childrenLength > 1) {
+    return false;
+  }
+
+  for (let i = 0; i < childrenLength; i++) {
+    const topBlock = children[i];
+
+    if ($isDecoratorNode(topBlock)) {
+      return false;
+    }
+
+    if ($isElementNode(topBlock)) {
+      if (!$isParagraphNode(topBlock)) {
+        return false;
+      }
+
+      if (topBlock.__indent !== 0) {
+        return false;
+      }
+
+      const topBlockChildren = topBlock.getChildren();
+      const topBlockChildrenLength = topBlockChildren.length;
+
+      for (let s = 0; s < topBlockChildrenLength; s++) {
+        const child = topBlockChildren[i];
+
+        if (!$isTextNode(child)) {
+          return false;
+        }
+      }
+    }
+  }
+
+  return true;
+}
+
+/**
+ * Returns a function that executes {@link $canShowPlaceholder}
+ * @param isEditorComposing - Is the editor in composition mode due to an active Input Method Editor?
+ * @returns A function that executes $canShowPlaceholder with arguments.
+ */
+export function $canShowPlaceholderCurry(
+  isEditorComposing: boolean,
+): () => boolean {
+  return () => $canShowPlaceholder(isEditorComposing);
+}
